* Moved panels into their own package so I can find them easier.
* Modify the TileSetCache, it is now a Singleton. #28

* Started the work on loading multiple TileSets from a board. #28
